docker: optionally skip upstream yum repo usage
This change introduces a conditional flag do_upstream_docker_repository
defaulting to true when using the docker role.

If set to true, the upstream yum repository will be used to install docker
packages and any related configuration will be performed (old behaviour).

If set to false, upstream repository will not be used, and the OS
provided docker package will be used.

Resolves: #688